GPayloadOptions, removed cmdUtils
